Check the error stack length before slicing
In Safari (iOS and desktop), the erorr stack never seems to be longer than
two frames, so when `arr.slice(2)` occurs, the result is an empty array, so
`arr[0]` is `undefined` rather than a string.

This checks the length of the array and only slices if there are at least 2
elements, otherwise, pass the array through wholesale.